refactor login token functionality
